Final Submisssion

+++ controller/comments.py
@@ -2,8 +2,8 @@ import sqlite3
 
 from flask_login import current_user
 
-from src.controller import error_printer
 from model.model import get_db
+from src.controller import error_printer
 
 
 def comment(pid, content):
@@ -11,7 +11,7 @@ def comment(pid, content):
     try:
         db.execute(f"insert into comments(user,post,content) values ('{current_user.email}','{pid}','{content}');")
         db.commit()
-        return 200
+        return "Comment created successfully", 200
     except sqlite3.Error as err:
         db.rollback()
         error_printer(err)
@@ -29,12 +29,12 @@ def get_comments(pid):
         return 406
 
 
-def edit_comment(pid, cid, content):
+def edit_comment(cid, content):
     db = get_db()
     try:
-        db.execute(f"update comments set content='{content}' where post='{pid}' and cid='{cid}'")
+        db.execute(f"update comments set content='{content}' where user='{current_user.email}' and  cid='{cid}'")
         db.commit()
-        return 200
+        return "Comment Updated Successfully", 200
     except sqlite3.Error as err:
         db.rollback()
         error_printer(err)
@@ -44,9 +44,9 @@ def edit_comment(pid, cid, content):
 def delete_comment(cid):
     db = get_db()
     try:
-        db.execute(f"delete from comments where cid='{cid}'")
+        db.execute(f"delete from comments where cid='{cid}' and user='{current_user.email}'")
         db.commit()
-        return 200
+        return "Comment deleted successfully", 200
     except sqlite3.Error as err:
         error_printer(err)
         return 406
